+++ backend/db/models/spot.js
@@ -21,12 +21,10 @@ module.exports = (sequelize, DataTypes) => {
 
       Spot.hasMany(models.Review, {
         foreignKey: 'spotId',
-        // as: 'avgRating'
       })
 
       Spot.hasMany(models.SpotImage, {
         foreignKey: 'spotId',
-        // as: 'previewImage'
       })
     }
   }
@@ -36,11 +34,11 @@ module.exports = (sequelize, DataTypes) => {
     city: DataTypes.STRING,
     state: DataTypes.STRING,
     country: DataTypes.STRING,
-    lat: DataTypes.DECIMAL,
-    lng: DataTypes.DECIMAL,
+    lat: DataTypes.FLOAT,
+    lng: DataTypes.FLOAT,
     name: DataTypes.STRING,
     description: DataTypes.STRING,
-    price: DataTypes.DECIMAL,
+    price: DataTypes.FLOAT,
     createdAt: DataTypes.DATE,
     updatedAt: DataTypes.DATE
   }, {
